# Compile `String.valid_utf8?/1` inline