@mtrudel 2023-01-19
Upon further review, this seems to actually be slower on x86. Rescinding this PR.