@josevalim 2023-01-19
Hi @mtrudel and @moogle19! Also, be sure to check OTP 26, it has some improvements around this area.

But I would also send a proposal to the Erlang/OTP team to add a `unicode:validate_utf8/1` function that is backed by a NIF. This is a very important operation and it can be highly optimised using techniques such as [simdutf8](https://github.com/rusticstuff/simdutf8). I would make it return something like `:ok` or `{:error, byte_pos_of_the_first_invalid_codepoint}`. I would love to see something like that get in, so let me know if you need help with any step of the process.