@josevalim 2023-01-19
I don't think you need an EEP for this. I would start a discussion on the forum or potentially just submit a PR depending on your appetite. :)

Pro-tip, in order to add NIFs to Erlang/OTP, first:

1. add the NIF stub to unicode.erl
2. them implement the nif

If you implement the nif before the stub, then you can't start the runtime because the stub the nif is supposed to fill in is missing. It may also be necessary to put the nif on a private module, such as erts_internal. I am not sure all modules from OTP are nif-able. :)