@mtrudel 2023-01-19
OK, this is *really* interesting.

Given this implementation:

```elixir
Mix.install([:benchee])

defmodule New do
  @compile {:inline, valid?: 1, valid_utf8?: 1}
  def valid?(<<string::binary>>), do: valid_utf8?(string)
  def valid?(_), do: false

  defp valid_utf8?(<<_::utf8, rest::bits>>), do: valid_utf8?(rest)
  defp valid_utf8?(<<>>), do: true
  defp valid_utf8?(_), do: false
end

defmodule New2 do
  @compile {:inline, valid?: 1, valid_utf8?: 1}
  def valid?(<<string::binary>>), do: valid_utf8?(string)
  def valid?(_), do: false

  defp valid_utf8?(<<a::8, b::8, c::8, d::8, rest::bits>>)
       when a < 128 and b < 128 and c < 128 and
              d < 128,
       do: valid_utf8?(rest)

  defp valid_utf8?(<<_::utf8, rest::bits>>), do: valid_utf8?(rest)
  defp valid_utf8?(<<>>), do: true
  defp valid_utf8?(_), do: false
end

Benchee.run(
  %{
    "old" => fn input -> String.valid?(input) end,
    "new" => fn input -> New.valid?(input) end,
    "new2" => fn input -> New2.valid?(input) end
  },
  time: 10,
  memory_time: 2,
  inputs: %{
    "micro" => String.duplicate("a", 10),
    "medium" => String.duplicate("a", 10_002)
  }
)
```

Here's what I get on OTP 25:

<details>
<summary>OTP 25 Benchmark</summary>
```
Operating System: macOS
CPU Information: Apple M1
Number of Available Cores: 8
Available memory: 16 GB
Elixir 1.14.1
Erlang 25.1

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 10 s
memory time: 2 s
reduction time: 0 ns
parallel: 1
inputs: medium, micro
Estimated total run time: 1.40 min

Benchmarking new with input medium ...
Benchmarking new with input micro ...
Benchmarking new2 with input medium ...
Benchmarking new2 with input micro ...
Benchmarking old with input medium ...
Benchmarking old with input micro ...

##### With input medium #####
Name           ips        average  deviation         median         99th %
new        37.94 K       26.36 μs    ±22.40%       25.92 μs          33 μs
old        33.67 K       29.70 μs    ±22.99%       29.29 μs       35.33 μs
new2       29.13 K       34.32 μs    ±30.66%       33.75 μs       41.58 μs

Comparison:
new        37.94 K
old        33.67 K - 1.13x slower +3.34 μs
new2       29.13 K - 1.30x slower +7.97 μs

Memory usage statistics:

Name    Memory usage
new            656 B
old            656 B - 1.00x memory usage +0 B
new2           656 B - 1.00x memory usage +0 B

**All measurements for memory usage were the same**

##### With input micro #####
Name           ips        average  deviation         median         99th %
new         2.91 M      343.35 ns ±11272.78%         250 ns         375 ns
old         2.81 M      355.54 ns ±11417.85%         250 ns         417 ns
new2        2.63 M      380.11 ns ±10750.79%         291 ns         458 ns

Comparison:
new         2.91 M
old         2.81 M - 1.04x slower +12.19 ns
new2        2.63 M - 1.11x slower +36.77 ns

Memory usage statistics:

Name    Memory usage
new            656 B
old            656 B - 1.00x memory usage +0 B
new2           656 B - 1.00x memory usage +0 B

**All measurements for memory usage were the same**
```
</details>

(tl;dr: inlining makes a little bit of difference, and fast-pathing ASCII strings with a 4 byte match actually slows things down)


Here's the exact same thing on OTP 26 (ref:master):

<details>
<summary>OTP 26 Benchmark</summary>
```
Operating System: macOS
CPU Information: Apple M1
Number of Available Cores: 8
Available memory: 16 GB
Elixir 1.14.1
Erlang 26.0-rc0

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 10 s
memory time: 2 s
reduction time: 0 ns
parallel: 1
inputs: medium, micro
Estimated total run time: 1.40 min

Benchmarking new with input medium ...
Benchmarking new with input micro ...
Benchmarking new2 with input medium ...
Benchmarking new2 with input micro ...
Benchmarking old with input medium ...
Benchmarking old with input micro ...

##### With input medium #####
Name           ips        average  deviation         median         99th %
new2      222.95 K        4.49 μs   ±232.18%        4.38 μs        4.67 μs
old        39.31 K       25.44 μs    ±25.66%       25.21 μs       26.21 μs
new        39.05 K       25.61 μs    ±30.18%       25.21 μs       32.63 μs

Comparison:
new2      222.95 K
old        39.31 K - 5.67x slower +20.95 μs
new        39.05 K - 5.71x slower +21.12 μs

Memory usage statistics:

Name    Memory usage
new2           656 B
old            656 B - 1.00x memory usage +0 B
new            656 B - 1.00x memory usage +0 B

**All measurements for memory usage were the same**

##### With input micro #####
Name           ips        average  deviation         median         99th %
new2        2.98 M      335.78 ns ±11879.12%         250 ns         416 ns
new         2.94 M      340.57 ns ±11380.92%         250 ns         375 ns
old         2.89 M      346.60 ns ±12061.87%         250 ns         416 ns

Comparison:
new2        2.98 M
new         2.94 M - 1.01x slower +4.79 ns
old         2.89 M - 1.03x slower +10.82 ns

Memory usage statistics:

Name    Memory usage
new2           656 B
new            656 B - 1.00x memory usage +0 B
old            656 B - 1.00x memory usage +0 B

**All measurements for memory usage were the same**
```

This time, we see basically no difference from inlining, but a *5.6x* difference on large ASCII strings by matching them 4 bytes at a time. WOW. 

On x86 the difference is smaller, at only 2x faster, but it's still twice as fast for a pretty small change.

The downside here is that this *exact* same change ends up 25% *slower* on OTP 25 and earlier. 

I'm going to spend a bit more time characterizing this but assuming it checks out, @josevalim is there precedent / best practices (or even an appetite) for per-OTP optimizations such as this in the standard library? Happy to work up a PR if so.