@josevalim 2023-01-19
How does it scale? Comparing to reading two/four/eight bytes at a time?